pyproject.toml
[tool.poetry]
name = "zas"
version = "0.1.0"
description = "Web site for Zulu-Alpha"
authors = ["Adam Piskorski <adam@piskorski.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4"
environs = { version = "*", extras = ["django"] }
psycopg2 = "^2.9.3"
django-allauth = "^0.47.0"
Jinja2 = "^3.0.3"
django-extensions = "^3.1.5"
django-debug-toolbar = "^3.2.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
mypy = "^0.910"
django-stubs = "^1.9.0"
flake8 = "^4.0.1"
flake8-bugbear = "^21.9.2"
pytest-django = "^4.4.0"
pytest-sugar = "^0.9.4"
pytest-xdist = "^2.4.0"
django-extensions = "^3.1.3"
bandit = "^1.7.1"
pre-commit = "^2.16.0"
rich = "^10.16.1"
black = "^22.1.0"
ipython = "^8.0.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 90
target-version = ['py310']

[tool.pytest.ini_options]
junit_family = "xunit1"
DJANGO_SETTINGS_MODULE = "config.settings"
addopts = "-n auto --nomigrations"
python_files = [
    "tests.py",
    "test_*.py",
    "tests_*.py",
    "*_tests.py",
    "*_test.py"
]
norecursedirs = [
    ".git",
    ".vscode",
    ".pytest_cache",
    ".mypy_cache",
    "__pycache__",
    "templates",
    "staticfiles",
    "migrations",
    "forms",
    "views",
    "models",
]
